Remove unused endpoint
